requirements.txt
openai>=1.0.0
jinja2>=3.1.2
pyyaml>=6.0
rdkit-pypi>=2022.9.5
numpy>=1.21
